backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
torch>=2.0.0
transformers>=4.28.0
Pillow>=9.5.0
pydantic>=1.10.7 